feat: 타입 분리, 페이지 분리
