+++ src/features/Login/Types.ts
@@ -1,0 +1,10 @@
+export interface LoginFormProps {
+  email: string;
+  password: string;
+}
+
+export interface TokenResponse {
+  item: {
+    token: string;
+  };
+}
